requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
openai==1.6.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
cryptography==41.0.7
python-dateutil==2.8.2